WRL Missions Brought to 10.7 Standards
I have spent hours making the WRL missions up to 10.7 standards. CTLD zones added to the miz files. Squadron limits added to the yamls. Most importantly the descriptions are all up to date and similar. Might be some minor touch ups from here on out but that majority of it is ready to go. I also deleted up the coast, retiring that legacy campaign which was my first release. Originally designed for liberation.