A-6A Intruder with Adecarcer's hardpoint options
Enabled Adecarcer's modified hardpoint options for the A-6A Intruder mod. This will, for instance, fix the MER bomb racks which have the "missing texture" material in the stock version of the mod. It will also allow mounting some weapons which are more appropriate for the A-6E variant of the aircraft. While not historically and technically accurate, this will allow employing the aircraft in a role which is closer to the EA-6B or A-6E, should the user wish to do so. The EA-6B mod is AI-only, so this will grant a player-flyable option. The use of custom loadouts is recommended.