WRL Faction Updates
I've updated all factions to have the same exact air assets to streamline the ability to use my factions anywhere. The difference in my factions is the units used in frontlines, SAMs, and so on. I still have to use custom factions to ensure my missions are played as I designed them. People can customize them from there. The new WRL mixed faction is going to be used in all wargame scenarios and some Sinai map stuff I've been looking at making. Also it really annoys me that the file name is not sticking. I want it to be caps or lowercase, not a mix.